Provides temp / salinity sampling for particles
Note, this is NOT OPTIMIZED FOR IO and these fields are always
communicated in the halo.

In order for this to be optimized these fields only need transfered for
writing out the fields, not for communication, which means they are
effetively nonhalo variables. Nonhalo variables were previously
deactivated for simplicity.