Fixes failed search for nearest hex vertex
Previously, nearest vertex selection approach failed because of
non-homogeneity in the mesh.  The existing approach takes a brute-force
approach and finds the mesh when neededing to do horizontal
interpolation.

Note, this is a far from optimal solution and it does not consider
corner cases where the vertex is on the boundary.